Removed unnecessary usings and separated end to end tests
