+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/CodeActions/CodeActionEndToEndTest.NetFx.cs
@@ -41,7 +41,7 @@ using Xunit.Sdk;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions;
 
-public class CodeActionEndToEndTest(ITestOutputHelper testOutput) : SingleServerDelegatingEndpointTestBase(testOutput)
+public partial class CodeActionEndToEndTest(ITestOutputHelper testOutput) : SingleServerDelegatingEndpointTestBase(testOutput)
 {
     private const string GenerateEventHandlerTitle = "Generate Event Handler 'DoesNotExist'";
     private const string ExtractToComponentTitle = "Extract element to new component";
@@ -65,28 +65,6 @@ public class CodeActionEndToEndTest(ITestOutputHelper testOutput) : SingleServer
                     new LspDocumentMappingService(FilePathService, new TestDocumentContextFactory(), LoggerFactory),
                     razorFormattingService)
             ];
-
-    private ExtractToComponentCodeActionResolver[] CreateExtractComponentCodeActionResolver(
-        string filePath,
-        RazorCodeDocument codeDocument,
-        IRazorFormattingService razorFormattingService,
-        IClientConnection clientConnection,
-        RazorLSPOptionsMonitor? optionsMonitor = null)
-    { 
-        var projectManager = new StrictMock<IDocumentVersionCache>();
-        int? version = 1;
-        projectManager.Setup(x => x.TryGetDocumentVersion(It.IsAny<IDocumentSnapshot>(), out version)).Returns(true);
-
-        return [
-                new ExtractToComponentCodeActionResolver(
-                    new GenerateMethodResolverDocumentContextFactory(filePath, codeDocument),
-                    optionsMonitor ?? TestRazorLSPOptionsMonitor.Create(),
-                    TestLanguageServerFeatureOptions.Instance,
-                    clientConnection,
-                    razorFormattingService,
-                    projectManager.Object)
-            ];
-    }
 
     #region CSharp CodeAction Tests
 
@@ -1034,314 +1012,6 @@ public class CodeActionEndToEndTest(ITestOutputHelper testOutput) : SingleServer
             diagnostics: [new Diagnostic() { Code = "CS0103", Message = "The name 'DoesNotExist' does not exist in the current context" }]);
     }
 
-    [Fact]
-    public async Task Handle_ExtractComponent_SingleElement_ReturnsResult()
-    {
-        var input = """
-            <[||]div id="a">
-                <h1>Div a title</h1>
-                <Book Title="To Kill a Mockingbird" Author="Harper Lee" Year="Long ago" />
-                <p>Div a par</p>
-            </div>
-            <div id="shouldSkip">
-                <Movie Title="Aftersun" Director="Charlotte Wells" Year="2022" />
-            </div>
-            """;
-
-        var expectedRazorComponent = """
-            <div id="a">
-                <h1>Div a title</h1>
-                <Book Title="To Kill a Mockingbird" Author="Harper Lee" Year="Long ago" />
-                <p>Div a par</p>
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_SiblingElement_ReturnsResult()
-    {
-        var input = """
-            <[|div id="a">
-                <h1>Div a title</h1>
-                <Book Title="To Kill a Mockingbird" Author="Harper Lee" Year="Long ago" />
-                <p>Div a par</p>
-            </div>
-            <div id="b">
-                <Movie Title="Aftersun" Director="Charlotte Wells" Year="2022" />
-            </div|]>
-            """;
-
-        var expectedRazorComponent = """
-            <div id="a">
-                <h1>Div a title</h1>
-                <Book Title="To Kill a Mockingbird" Author="Harper Lee" Year="Long ago" />
-                <p>Div a par</p>
-            </div>
-            <div id="b">
-                <Movie Title="Aftersun" Director="Charlotte Wells" Year="2022" />
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_StartNodeContainsEndNode_ReturnsResult()
-    {
-        var input = """
-            <[|div id="parent">
-                <div>
-                    <div>
-                        <div>
-                            <p>Deeply nested par</p|]>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            """;
-
-        var expectedRazorComponent = """
-            <div id="parent">
-                <div>
-                    <div>
-                        <div>
-                            <p>Deeply nested par</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_SiblingElement_ReturnsResult()
-    {
-        var input = """
-            <[|div id="a">
-                <h1>Div a title</h1>
-                <Book Title="To Kill a Mockingbird" Author="Harper Lee" Year="Long ago" />
-                <p>Div a par</p>
-            </div>
-            <div id="b">
-                <Movie Title="Aftersun" Director="Charlotte Wells" Year="2022" />
-            </div|]>
-            """;
-
-        var expectedRazorComponent = """
-            <div id="a">
-                <h1>Div a title</h1>
-                <Book Title="To Kill a Mockingbird" Author="Harper Lee" Year="Long ago" />
-                <p>Div a par</p>
-            </div>
-            <div id="b">
-                <Movie Title="Aftersun" Director="Charlotte Wells" Year="2022" />
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_StartNodeContainsEndNode_ReturnsResult()
-    {
-        var input = """
-            <[|div id="parent">
-                <div>
-                    <div>
-                        <div>
-                            <p>Deeply nested par</p|]>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            """;
-
-        var expectedRazorComponent = """
-            <div id="parent">
-                <div>
-                    <div>
-                        <div>
-                            <p>Deeply nested par</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_IndentedNode_ReturnsResult()
-    {
-        var input = """
-            <div id="parent">
-                <div>
-                    <[||]div>
-                        <div>
-                            <p>Deeply nested par</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            """;
-
-        var expectedRazorComponent = """
-            <div>
-                <div>
-                    <p>Deeply nested par</p>
-                </div>
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_IndentedSiblingNodes_ReturnsResult()
-    {
-        var input = """
-            <div id="parent">
-                <div>
-                    <div>
-                        <div>
-                            <[|div>
-                                <p>Deeply nested par</p>
-                            </div>
-                            <div>
-                                <p>Deeply nested par</p>
-                            </div>
-                            <div>
-                                <p>Deeply nested par</p>
-                            </div>
-                            <div>
-                                <p>Deeply nested par</p>
-                            </div|]>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            """;
-
-        var expectedRazorComponent = """
-            <div>
-                <p>Deeply nested par</p>
-            </div>
-            <div>
-                <p>Deeply nested par</p>
-            </div>
-            <div>
-                <p>Deeply nested par</p>
-            </div>
-            <div>
-                <p>Deeply nested par</p>
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_IndentedStartNodeContainsEndNode_ReturnsResult()
-    {
-        var input = """
-            <div id="parent">
-                <div>
-                    <[|div>
-                        <div>
-                            <p>Deeply nested par</p|]>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            """;
-
-        var expectedRazorComponent = """
-            <div>
-                <div>
-                    <p>Deeply nested par</p>
-                </div>
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
-    [Fact]
-    public async Task Handle_ExtractComponent_IndentedEndNodeContainsStartNode_ReturnsResult()
-    {
-        var input = """
-            <div id="parent">
-                <div>
-                    <div>
-                        <div>
-                            <[|p>Deeply nested par</p>
-                        </div>
-                    </div|]>
-                </div>
-            </div>
-            """;
-
-        var expectedRazorComponent = """
-            <div>
-                <div>
-                    <p>Deeply nested par</p>
-                </div>
-            </div>
-            """;
-
-        await ValidateExtractComponentCodeActionAsync(
-            input,
-            expectedRazorComponent,
-            ExtractToComponentTitle,
-            razorCodeActionProviders: [new ExtractToComponentCodeActionProvider(LoggerFactory)],
-            codeActionResolversCreator: CreateExtractComponentCodeActionResolver);
-    }
-
     #endregion
 
     private async Task ValidateCodeBehindFileAsync(
@@ -1482,61 +1152,6 @@ public class CodeActionEndToEndTest(ITestOutputHelper testOutput) : SingleServer
         }
 
         var actual = sourceText.WithChanges(edits).ToString();
-        AssertEx.EqualOrDiff(expected, actual);
-    }
-
-    private async Task ValidateExtractComponentCodeActionAsync(
-        string input,
-        string? expected,
-        string codeAction,
-        int childActionIndex = 0,
-        IEnumerable<(string filePath, string contents)>? additionalRazorDocuments = null,
-        IRazorCodeActionProvider[]? razorCodeActionProviders = null,
-        Func<string, RazorCodeDocument, IRazorFormattingService, IClientConnection, RazorLSPOptionsMonitor?, IRazorCodeActionResolver[]>? codeActionResolversCreator = null,
-        RazorLSPOptionsMonitor? optionsMonitor = null,
-        Diagnostic[]? diagnostics = null)
-    {
-        TestFileMarkupParser.GetSpan(input, out input, out var textSpan);
-
-        var razorFilePath = "C:/path/to/test.razor";
-        var componentFilePath = "C:/path/to/Component.razor";
-        var codeDocument = CreateCodeDocument(input, filePath: razorFilePath);
-        var sourceText = codeDocument.Source.Text;
-        var uri = new Uri(razorFilePath);
-        var languageServer = await CreateLanguageServerAsync(codeDocument, razorFilePath, additionalRazorDocuments);
-        var documentContext = CreateDocumentContext(uri, codeDocument);
-        var requestContext = new RazorRequestContext(documentContext, null!, "lsp/method", uri: null);
-
-        var result = await GetCodeActionsAsync(
-            uri,
-            textSpan,
-            sourceText,
-            requestContext,
-            languageServer,
-            razorCodeActionProviders,
-            diagnostics);
-
-        Assert.NotEmpty(result);
-        var codeActionToRun = GetCodeActionToRun(codeAction, childActionIndex, result);
-
-        if (expected is null)
-        {
-            Assert.Null(codeActionToRun);
-            return;
-        }
-
-        Assert.NotNull(codeActionToRun);
-
-        var formattingService = await TestRazorFormattingService.CreateWithFullSupportAsync(LoggerFactory, codeDocument, documentContext.Snapshot, optionsMonitor?.CurrentValue);
-        var changes = await GetEditsAsync(
-            codeActionToRun,
-            requestContext,
-            languageServer,
-            codeActionResolversCreator?.Invoke(razorFilePath, codeDocument, formattingService, languageServer, arg5: null) ?? []);
-
-        var edits = changes.Where(change => change.TextDocument.Uri.AbsolutePath == componentFilePath).Single();
-        var actual = edits.Edits.Select(edit => edit.NewText).Single();
-
         AssertEx.EqualOrDiff(expected, actual);
     }
 
@@ -1731,80 +1346,6 @@ public class CodeActionEndToEndTest(ITestOutputHelper testOutput) : SingleServer
 
             var projectWorkspaceState = ProjectWorkspaceState.Create(_tagHelperDescriptors.ToImmutableArray());
             var testDocumentSnapshot = TestDocumentSnapshot.Create(FilePath, CodeDocument.Source.Text.ToString(), CodeAnalysis.VersionStamp.Default, projectWorkspaceState);
-            testDocumentSnapshot.With(CodeDocument);
-
-            context = CreateDocumentContext(new Uri(FilePath), testDocumentSnapshot);
-            return true;
-        }
-
-        private static List<TagHelperDescriptor> CreateTagHelperDescriptors()
-        {
-            return BuildTagHelpers().ToList();
-
-            static IEnumerable<TagHelperDescriptor> BuildTagHelpers()
-            {
-                var builder = TagHelperDescriptorBuilder.Create("oncontextmenu", "Microsoft.AspNetCore.Components");
-                builder.SetMetadata(
-                    new KeyValuePair<string, string>(ComponentMetadata.EventHandler.EventArgsType, "Microsoft.AspNetCore.Components.Web.MouseEventArgs"),
-                    new KeyValuePair<string, string>(ComponentMetadata.SpecialKindKey, ComponentMetadata.EventHandler.TagHelperKind));
-                yield return builder.Build();
-
-                builder = TagHelperDescriptorBuilder.Create("onclick", "Microsoft.AspNetCore.Components");
-                builder.SetMetadata(
-                    new KeyValuePair<string, string>(ComponentMetadata.EventHandler.EventArgsType, "Microsoft.AspNetCore.Components.Web.MouseEventArgs"),
-                    new KeyValuePair<string, string>(ComponentMetadata.SpecialKindKey, ComponentMetadata.EventHandler.TagHelperKind));
-
-                yield return builder.Build();
-
-                builder = TagHelperDescriptorBuilder.Create("oncopy", "Microsoft.AspNetCore.Components");
-                builder.SetMetadata(
-                    new KeyValuePair<string, string>(ComponentMetadata.EventHandler.EventArgsType, "Microsoft.AspNetCore.Components.Web.ClipboardEventArgs"),
-                    new KeyValuePair<string, string>(ComponentMetadata.SpecialKindKey, ComponentMetadata.EventHandler.TagHelperKind));
-
-                yield return builder.Build();
-
-                builder = TagHelperDescriptorBuilder.Create("ref", "Microsoft.AspNetCore.Components");
-                builder.SetMetadata(
-                    new KeyValuePair<string, string>(ComponentMetadata.SpecialKindKey, ComponentMetadata.Ref.TagHelperKind),
-                    new KeyValuePair<string, string>(ComponentMetadata.Common.DirectiveAttribute, bool.TrueString));
-
-                yield return builder.Build();
-            }
-        }
-    }
-
-    private class ExtractToComponentResolverDocumentContextFactory : TestDocumentContextFactory
-    {
-        private readonly List<TagHelperDescriptor> _tagHelperDescriptors;
-
-        public ExtractToComponentResolverDocumentContextFactory
-            (string filePath,
-            RazorCodeDocument codeDocument,
-            TagHelperDescriptor[]? tagHelpers = null,
-            int? version = null)
-            : base(filePath, codeDocument, version)
-        {
-            _tagHelperDescriptors = CreateTagHelperDescriptors();
-            if (tagHelpers is not null)
-            {
-                _tagHelperDescriptors.AddRange(tagHelpers);
-            }
-        }
-
-        public override bool TryCreate(
-            Uri documentUri,
-            VSProjectContext? projectContext,
-            bool versioned,
-            [NotNullWhen(true)] out DocumentContext? context)
-        {
-            if (FilePath is null || CodeDocument is null)
-            {
-                context = null;
-                return false;
-            }
-
-            var projectWorkspaceState = ProjectWorkspaceState.Create(_tagHelperDescriptors.ToImmutableArray());
-            var testDocumentSnapshot = TestDocumentSnapshot.Create(FilePath, CodeDocument.GetSourceText().ToString(), CodeAnalysis.VersionStamp.Default, projectWorkspaceState);
             testDocumentSnapshot.With(CodeDocument);
 
             context = CreateDocumentContext(new Uri(FilePath), testDocumentSnapshot);
